libdxfrw: hygiene — EOF newlines, seed comments, redundant cast/(void)
Low-risk cleanups from the review's style/dead-code findings:
- Restore the missing trailing newline at EOF in drw_base.h, drw_classes.h,
  drw_interface.h.
- Correct seedReservedDxf comments: 0x1C is BLOCK "*Paper_Space" and 0x1D is
  ENDBLK "*Paper_Space" (verified against the codec's own emission at the
  writeString(5,"1C")/("1D") sites), not the stale ENDBLK/"(legacy)" labels.
- Drop the redundant (void)before in drw_header.cpp ($HANDSEED back-patch):
  `before` is used at the guard above, so the suppression is dead.
- Drop the redundant static_cast<uint32_t> in DRW_Ole2Frame::parseDwg: the
  operand remainingBytes is already const std::uint32_t.

Two cleanups the review proposed were verified NOT redundant and deliberately
left alone: the INSERT code-66 writes live in distinct version branches (each
emits the attributes-follow flag for its own DXF version), and the
dwgreader18.cpp setBitPos(0) is functional byte-alignment (setPosition does
not reset the bit position, per resetPosition's separate setBitPos(0) call).

+++ libraries/libdxfrw/src/drw_entities.cpp
@@ -2908,7 +2908,7 @@ bool DRW_Ole2Frame::parseDwg(DRW::Version v, dwgBuffer *buf, std::uint32_t bs){
         m_payloadTooLarge = true;
         return false;
     }
-    if (m_declaredPayloadLength > static_cast<std::uint32_t>(remainingBytes)) {
+    if (m_declaredPayloadLength > remainingBytes) {  // remainingBytes is already uint32
         m_payloadTruncated = true;
         m_payloadByteCount = remainingBytes;
         return false;
